bitcoin_wallet_app/requirements.txt
PyQt5
python-bitcoinlib
bip_utils
